Refactor backend structure and add search app to align with RESTful APIs
