+++ backend/projects/urls.py
@@ -5,7 +5,6 @@ from rest_framework.urlpatterns import format_suffix_patterns
 from . import views
 
 urlpatterns = [
-    path("search/", views.ProjectSearchView.as_view()),
     path("reorder/", views.ProjectReorderView.as_view()),
     path("", views.ProjectList.as_view()),
     path("inbox/", views.InboxProjectDetail.as_view()),
