+++ backend/projects/views.py
@@ -114,43 +114,3 @@ class ProjectReorderView(mixins.UpdateModelMixin, generics.GenericAPIView):
         Project.objects.bulk_update(projects, ["order"])
 
         return Response(status=status.HTTP_200_OK)
-
-class ProjectSearchPagination(PageNumberPagination):
-    page_size = 20
-
-# 하나의 view에서 여러 serializer를 부를 거라 복잡해도 그냥 APIView가 적합할 거 같음
-# 파라미터 종류가 많아지면 django-filter를 적극적으로 고려해 보자...
-class ProjectSearchView(APIView):
-    def get(self, request, *args, **kwargs):
-        q = request.query_params.get("query", "").strip()
-
-        # query가 비어있으면 빈 결과 반환
-        # TODO: 나중에 프로젝트 페이지와 합치게 되면 전부 보이는 걸로 바뀌어야 할 지도?
-        if not q:
-            return Response({
-                "projects": [],
-                "drawers": [],
-                "tasks": [],
-            })
-
-        project_qs = Project.objects.filter(
-            Q(name__icontains=q)
-        )
-
-        drawer_qs = Drawer.objects.filter(
-            Q(name__icontains=q)
-        )
-
-        task_qs = Task.objects.filter(
-            Q(name__icontains=q)
-        )
-
-        project_data = ProjectSearchSerializer(project_qs, many=True).data
-        drawer_data = DrawerSearchSerializer(drawer_qs, many=True).data
-        task_data = TaskSearchSerializer(task_qs, many=True).data
-
-        return Response({
-            "projects": project_data,
-            "drawers": drawer_data,
-            "tasks": task_data,
-        })
