+++ backend/search/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+
+from rest_framework.urlpatterns import format_suffix_patterns
+
+from . import views
+
+urlpatterns = [
+    path("", views.ProjectSearchView.as_view()),
+]
+
+urlpatterns = format_suffix_patterns(urlpatterns)
